requirements.txt
django
gunicorn
djangorestframework
drf-extensions
numpy
opencv-python
pillow
django-cors-headers
django-heroku
django-filter
rq
redis
celery
jsonfield
boto3
boto
django-storages
django-storages-redux
